requirements.txt
# Base de datos y ORM
sqlalchemy[asyncio]~=2.0.36
alembic~=1.14.0
asyncpg~=0.30.0
psycopg[binary]~=3.2.3
GeoAlchemy2~=0.15.2

# GraphQL
strawberry-graphql[asgi]~=0.243.0
strawberry-sqlalchemy-mapper~=0.7.0
aiodataloader~=0.3.0

# Servidor web
uvicorn[standard]~=0.32.1
starlette~=0.41.3

# Utilidades
pandas>=2.2.0,<3.0.0
requests~=2.32.0
python-dotenv~=1.0.1
unidecode~=1.3.8